fix: fix sprintf_s type mismatch and ambiguity in core_types.h and chart_editor_settings.cpp

+++ src/peepodrumkit/chart_editor_settings.cpp
@@ -55,10 +55,10 @@ namespace PeepoDrumKit
 		return false;
 	}
 
-	static i32 RectToTLSizeString(char* buffer, size_t bufferSize, const Rect& in)
-	{
-		return sprintf_s(buffer, bufferSize, "%g, %g, %g, %g", in.TL.x, in.TL.y, in.GetWidth(), in.GetHeight());
-	}
+		static i32 RectToTLSizeString(char* buffer, size_t bufferSize, const Rect& in)
+		{
+			return sprintf_s(buffer, (size_t)bufferSize, "%g, %g, %g, %g", in.TL.x, in.TL.y, in.GetWidth(), in.GetHeight());
+		}
 
 	static b8 RectFromTLSizeString(std::string_view in, Rect& out)
 	{
